Extract the rest of the consts
