+++ src/utils/consts/ListName.ts
@@ -1,0 +1,1 @@
+export type ListName = 'storage' | 'shoppingList';
