+++ src/utils/consts/monthList.ts
@@ -1,0 +1,14 @@
+export const monthList = [
+  'January',
+  'February',
+  'March',
+  'April',
+  'May',
+  'June',
+  'July',
+  'August',
+  'September',
+  'October',
+  'November',
+  'December'
+];
