+++ src/utils/consts/pages.ts
@@ -1,0 +1,14 @@
+export const pages = {
+  storage: {
+    default: require('@/assets/images/Empty.svg'),
+    selected: require('@/assets/images/Check.svg')
+  },
+  newProduct: {
+    default: require('@/assets/images/PlusIcon.svg'),
+    selected: require('@/assets/images/Minus.svg')
+  },
+  shoppingList: {
+    default: require('@/assets/images/Empty.svg'),
+    selected: require('@/assets/images/Check.svg')
+  }
+};
